+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Future changes will appear here.
 
 
 
+
+## [1.5.17] - 2025-09-22
+
+### Features
+
+- add precise token precision conversion functions ([#475](https://github.com/autonomys/auto-sdk/pull/475)) [@jfrank-summit](https://github.com/jfrank-summit)
+
+### Chores
+
+- v1.5.16 ([#473](https://github.com/autonomys/auto-sdk/pull/473)) [@jfrank-summit](https://github.com/jfrank-summit)
+
 ## [1.5.16] - 2025-09-18
 
 ### Features
@@ -33,6 +44,7 @@ Future changes will appear here.
 
 - v1.5.14 ([#465](https://github.com/autonomys/auto-sdk/pull/465)) [@jfrank-summit](https://github.com/jfrank-summit)
 
+
 ## [1.5.14] - 2025-08-08
 
 ### Features
@@ -42,6 +54,7 @@ Future changes will appear here.
 ### Chores
 
 - v1.5.13 ([#461](https://github.com/autonomys/auto-sdk/pull/461)) [@jfrank-summit](https://github.com/jfrank-summit)
+
 
 
 ## [1.5.13] - 2025-08-08
@@ -57,6 +70,7 @@ Future changes will appear here.
 ### Chores
 
 - update dependency @polkadot/extension-inject to ^0.58.9 ([#358](https://github.com/autonomys/auto-sdk/pull/358)) [@renovate[bot]](https://github.com/apps/renovate)
+
 
 
 
@@ -76,11 +90,13 @@ Future changes will appear here.
 
 
 
+
 ## [1.5.11] - 2025-07-21
 
 ### Features
 
 - add get node method ([#454](https://github.com/autonomys/auto-sdk/pull/454)) [@clostao](https://github.com/clostao)
+
 
 
 
@@ -96,11 +112,13 @@ Future changes will appear here.
 
 
 
+
 ## [1.5.9] - 2025-07-15
 
 ### Features
 
 - simplify `nominatorPosition` to use runtime api ([#450](https://github.com/autonomys/auto-sdk/pull/450)) [@jfrank-summit](https://github.com/jfrank-summit)
+
 
 
 
@@ -120,11 +138,13 @@ Future changes will appear here.
 
 
 
+
 ## [1.5.7] - 2025-07-09
 
 ### Code Refactoring
 
 - update sdk structure for avoid deps issues ([#443](https://github.com/autonomys/auto-sdk/pull/443)) [@clostao](https://github.com/clostao)
+
 
 
 
@@ -140,12 +160,14 @@ Future changes will appear here.
 
 
 
+
 ## [1.5.5] - 2025-07-02
 
 ### Features
 
 - add headDomainNumber function to retrieve latest block number for a specific domain ([#440](https://github.com/autonomys/auto-sdk/pull/440)) [@jfrank-summit](https://github.com/jfrank-summit)
 - add storage fee refund to pending withdrawals ([#439](https://github.com/autonomys/auto-sdk/pull/439)) [@jfrank-summit](https://github.com/jfrank-summit)
+
 
 
 
@@ -158,6 +180,7 @@ Future changes will appear here.
 ### Chores
 
 - v1.5.3 ([#437](https://github.com/autonomys/auto-sdk/pull/437)) [@clostao](https://github.com/clostao)
+
 
 
 
@@ -183,6 +206,7 @@ Future changes will appear here.
 
 
 
+
 ## [1.5.2] - 2025-06-24
 
 ### Features
@@ -195,6 +219,7 @@ Future changes will appear here.
 
 
 
+
 ## [1.5.1] - 2025-06-19
 
 ### Features
@@ -204,6 +229,7 @@ Future changes will appear here.
 ### Chores
 
 - bump to v1.5.0 ([#426](https://github.com/autonomys/auto-sdk/pull/426)) [@clostao](https://github.com/clostao)
+
 
 
 
@@ -225,6 +251,7 @@ Future changes will appear here.
 
 
 
+
 ## [1.4.35] - 2025-05-29
 
 ### Features
@@ -234,6 +261,7 @@ Future changes will appear here.
 ### Chores
 
 - bump to v1.4.34 ([#420](https://github.com/autonomys/auto-sdk/pull/420)) [@clostao](https://github.com/clostao)
+
 
 
 
@@ -249,6 +277,7 @@ Future changes will appear here.
 
 
 
+
 ## [1.4.33] - 2025-05-28
 
 ### Bug Fixes
@@ -258,6 +287,7 @@ Future changes will appear here.
 ### Chores
 
 - v1.4.32 ([#417](https://github.com/autonomys/auto-sdk/pull/417)) [@clostao](https://github.com/clostao)
+
 
 
 
@@ -377,6 +407,7 @@ Future changes will appear here.
 
 
 
+
 ## [1.4.31] - 2025-05-14
 
 ### Features
@@ -393,6 +424,8 @@ Future changes will appear here.
 
 [1.5.14]: https://github.com/autonomys/auto-sdk/releases/tag/v1.5.14
 
-[Unreleased]: https://github.com/autonomys/auto-sdk/compare/v1.5.16...HEAD
-[1.5.16]: https://github.com/autonomys/auto-sdk/compare/v1.5.15...v1.5.16
 [1.5.15]: https://github.com/autonomys/auto-sdk/releases/tag/v1.5.15
+
+[Unreleased]: https://github.com/autonomys/auto-sdk/compare/v1.5.17...HEAD
+[1.5.17]: https://github.com/autonomys/auto-sdk/compare/v1.5.16...v1.5.17
+[1.5.16]: https://github.com/autonomys/auto-sdk/releases/tag/v1.5.16